# landing-page/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GmailMCP API",
    "version": "0.1.0",
    "description": "GmailMCP is a Model Context Protocol server that gives your AI agent real capabilities: one codebase exposed over CLI, MCP, and HTTP."
  },
  "paths": {
...
    "/api/v1/agentic-payments/requirements": {
      "post": {
        "tags": [
          "agentic-payments"
        ],
        "summary": "Create Payment Requirement",
        "description": "Generate a payment requirement for a given protocol.\n\nReturns the data needed to construct a 402 response.",
        "operationId": "create_payment_requirement_api_v1_agentic_payments_requirements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequirementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Payment Requirement Api V1 Agentic Payments Requirements Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agentic-payments/verify": {
      "post": {
        "tags": [
          "agentic-payments"
        ],
        "summary": "Verify Payment",
        "description": "Verify an incoming payment payload against a requirement.",
        "operationId": "verify_payment_api_v1_agentic_payments_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Verify Payment Api V1 Agentic Payments Verify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PaymentRequirementRequest": {
        "properties": {
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "default": "x402"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "asset": {
            "type": "string",
            "title": "Asset",
            "default": "USDC"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient",
            "default": ""
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "PaymentRequirementRequest",
        "description": "Request body for creating a payment requirement."
      },
      "PaymentVerifyRequest": {
        "properties": {
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "default": "x402"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "header_value": {
            "type": "string",
            "title": "Header Value",
            "default": ""
          },
          "requirement": {
            "additionalProperties": true,
            "type": "object",
            "title": "Requirement"
          }
        },
        "type": "object",
        "required": [
          "payload",
          "requirement"
        ],
        "title": "PaymentVerifyRequest",
        "description": "Request body for verifying a payment payload."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
  "servers": [
    {
      "url": "https://api.gmailmcp.com",
      "description": "GmailMCP production API"
    }
  ]
}